created SceneManager class, initializing spinners and combo boxes

+++ src/model/Administrator.java
@@ -13,7 +13,7 @@ public class Administrator extends Freelancer{
         this.auditorNBU = auditorNBU;
     }
 
-    public Administrator(String name, double dailyWage, int experienceYears, Freelancer.Education highestEducation, ArrayList<String> certificates, boolean auditorNBU) {
+    public Administrator(String name, int dailyWage, int experienceYears, Freelancer.Education highestEducation, ArrayList<String> certificates, boolean auditorNBU) {
         this.setName(name);
         this.setDailyWage(dailyWage);
         this.setExperienceYears(experienceYears);
